64bit

+++ CKTransferRecord.h
@@ -42,7 +42,7 @@
 	NSTimeInterval _lastTransferTime;
 	NSTimeInterval _transferStartTime;
 	NSTimeInterval _lastDirectorySpeedUpdate;
-	float _speed;
+	CGFloat _speed;
 	NSUInteger _progress;
 	NSMutableArray *_contents;
 	CKTransferRecord *_parent; //not retained
@@ -61,8 +61,8 @@
 - (unsigned long long)size;
 - (void)setSize:(unsigned long long)size;
 
-- (float)speed;
-- (void)setSpeed:(float)speed;	// TODO: Switch to CGFloat
+- (CGFloat)speed;
+- (void)setSpeed:(CGFloat)speed;	// TODO: Switch to CGFloat
 
 - (NSError *)error;
 
@@ -106,7 +106,7 @@
 // If the path is absolute, searches from root of tree, otherwise searches from receiver
 - (CKTransferRecord *)recordForPath:(NSString *)path;
 
-- (BOOL)problemsTransferringCountingErrors:(int *)outErrors successes:(int *)outSuccesses;
+- (BOOL)problemsTransferringCountingErrors:(NSInteger *)outErrors successes:(NSInteger *)outSuccesses;
 
 @end
